Add bin/check-framework-defaults
Needs a refactor to either extract common logic with
bin/check-config-docs or combine the two scripts but for now this is
working.

There's also a TODO for better handling HashLiterals and multi-line
strings as the values of configuration options. Currently it just grabs
the existing value from the doc but that obviously will not work when
trying to generate values that are not in the doc