Replace FileEditor with CachedParser and Resolver
The cached parser will be used to pass ASTs to the check classes. Since
checks will likely need access to the same ASTs, the parsing is cached.

The interesting functionality of the RailsFileEditor was that it
used the rails_path to get absolute paths from relative paths. That part
was extracted into its own class, and regular File operations are now
using the pure File methods.