Add assert_offense helper
This badly copies the idea of Rubocop's expect_offense. The gist is the
inline offense annotation gets parsed and then the attributes of the
parsed annotation become the expected values for the remaining lines'
offense.

Adding some more unit tests also revealed a discrepency with the ranges
for some offenses. The reason these worked before is that the count of
the range would be larger than the end, so the rjust was a noop. Adding
assertions for the exact ranges revealed this inconsistency and it is
now fixed.