Improve probably_header? heuristic
The Action Pack changelog recently ended up in a state where new entries
would fail check-changelogs even though the new entries were formatted
correctly. The issue is that probably_header? tries to guess whether
something is a header or an author, and tries to take into account bad
formatting. In this case, a header line ended with a * character, and
probably_header? saw that and assumed that it must be a badly formatted
author line.

This change tries to improve how probably_header? determines whether
something is a header. In all correctly formatted author lines, there
should be an even number of * characters because they surround author
names to make them bold. Therefore, we can make probably_header? return
true for lines that contain an odd number of * characters.

This is not perfect, as a badly formatted author line could contain an
extraneous *. However, it does fix the problem mentioned above, and it
should be correct in the majority of cases.

Maybe an ideal solution would be to weight a couple of different things
like this to try to make a more accurate guess? On one hand, that could
probably make the guessing even more accurate, but on the other hand
this is the first misprediction so far.