Extract TargetVersionCaseFinder from Visitor
This is the first in a series of refactors to improve how
Visitor::FrameworkDefault works. Overall, the goal is to componentize
each part of the logic and ideally prevent errors when new code patterns
are added to Rails::Application::Configuration.

Previously, Visitor::FrameworkDefault used visit_case and an instance
variable to control when the other visitor methods would run.

While this worked fine, it makes more sense to split up the two parts:
- TargetVersionCaseFinder is extracted as a SyntaxTree::Search, which
  performs a BFS on the AST to find the target_version case node
- the FrameworkDefault visitor can then be simplified since it will only
  be operating within the case statement and no longer has to worry
  about any nodes outside of it