Fix if statements inside respond_to? raising
Previously, any if statement inside a respond_to?(:framework) check
would clobber the @current_framework and cause assert_framework to raise
an error.

This refactors the VersionedConfig visitor to be smarter when visiting
if statements. The @current_framework has been changed to a stack in
case there end up being nested respond_to? checks (and STRICT=true will
make them raise). In addition, if statements that are not respond_to?
checks will now be ignored so that they do not cause assert_framework to
raise.