Extract visit_when from FrameworkDefaultInternal
This is part of a series of refactors to improve how
Visitor::FrameworkDefault works. Overall, the goal is to componentize
each part of the logic and ideally prevent errors when new code patterns
are added to Rails::Application::Configuration.

Previously, the FrameworkDefaultInternal Visitor handled visiting the
case node's when nodes. It accomplished this by setting an instance
variable when it sees a when node to the node's argument (the version).

The new approach iterates over the when nodes explicitly, instead of
visiting them in the Internal Visitor. This will preemptively prevent
issues if there ends up being nested case statements inside the
existing case statement (however unlikely that is).